Update pull request policy
Contributors should not send pull requests to the master branch;
they should target their PRs towards the appropriate branch in
order to prevent breaking the production system.

A more detailed branch catalogue and corresponding documentation
should be worked on in the future, in order to facilitate the
efforts of contributors.