Relicense under MIT/X11 license
In order to support inclusion of this project in others, we are
switching from the GNU Affero General Public License 3.0 to the
MIT/X11 license. We can only perform this transition under the
condition that the copyright holders all are in agreement of the
transition at time of the change, and since Carter Hinsley is the
current only copyright holder, this transition is valid and so the
project is now MIT/X11 licensed. :)